rename to SetMarketMetadata
